fixed seed issues

+++ db/migrate/20191113101032_create_locations.rb
@@ -3,8 +3,8 @@ class CreateLocations < ActiveRecord::Migration[6.0]
     create_table :locations do |t|
       t.string :area
       t.string :description
-      t.decimal :latitude
-      t.decimal :longitude
+      t.float :latitude
+      t.float :longitude
       t.integer :user_id
 
       t.timestamps
